Robust DCF77 signal decoding library.
Expects to be given 32 equally-spaced samples per second (e.g. using a timer interrupt), giving true for the unmodulated signal and false for the negatively-modulated pulse signal.

Pulse edge detection is based on a 15-second running average for each of the 32 sample points per second, which makes it robust against irregular sampling errors. Pulse length detection counts bits and uses thresholds to determine whether a pulse is short (signifying a 0-bit in the time signal) or long (a 1-bit).

Once a complete set of time signal bits has been gathered, the time and date is decoded, validated and passed to a callback function.

Uses 101 bytes of SRAM runtime state.